+++ src/main/java/br/com/ttrans/samapp/dao/EventDao.java
@@ -10,7 +10,7 @@ public interface EventDao {
 	public void add(Event event);
 	public void edit(Event event, Authentication authentication);
 	public void delete(Event event, Authentication authentication);
-	public Event get(int id); 
+	public Event get(long id); 
 	public List getAll();
 	public List loadData();
 }
